Added ANT build, added actual libs used to project
If I can't .gitignore the .classpath file I might as well add the libs
to the project, and make it all a lot cleaner.

The ant build script will make a .jar file and delete the ./bin/
directory as it's highly irrelevant after the .jar is complete.

All this is fairly untested at this point, but as the plugin has a
total of ONE user, myself, I don't see a problem.